+++ src/xml2_node.cpp
@@ -295,7 +295,15 @@ void node_set_attr(XPtrNode node_, std::string name, SEXP value, CharacterVector
   }
   if (hasPrefix("xmlns:", name)) {
     std::string prefix = name.substr(6);
-    xmlAddNamespace(node, xmlNewNs(node, asXmlChar(value), asXmlChar(prefix)));
+    xmlNsPtr ns = xmlSearchNs(node_->doc, node, asXmlChar(prefix));
+
+    if (ns == NULL) {
+      xmlAddNamespace(node, xmlNewNs(node, asXmlChar(value), asXmlChar(prefix)));
+    } else {
+      ns->href = xmlStrdup(asXmlChar(value));
+      xmlSetNs(node, ns);
+    }
+
     return;
   }